light, medium, hard works

+++ app/src/main/java/com/example/myverysmarthome/model/devices/Thermostat.java
@@ -2,8 +2,8 @@ package com.example.myverysmarthome.model.devices;
 
 import com.example.myverysmarthome.R;
 
-public class Thermostat extends Device<Integer> {
+public class Thermostat extends Device<Float> {
     public Thermostat(String name) {
-        super(name, 23, R.drawable.temperature);
+        super(name, 23.0f, R.drawable.temperature);
     }
 }
